Introduce AbstractMethod and TraitRefrence.
AbstractMethod is the base class for Method and Trait.
