+++ lib/raml.rb
@@ -3,6 +3,7 @@ require_relative 'raml/version'
 require_relative 'raml/common'
 
 require_relative 'raml/parameter/abstract_parameter'
+require_relative 'raml/abstract_method'
 
 require_relative 'raml/protocol'
 require_relative 'raml/method'
@@ -13,6 +14,7 @@ require_relative 'raml/response'
 require_relative 'raml/body'
 require_relative 'raml/header'
 require_relative 'raml/documentation'
+require_relative 'raml/trait_reference'
 require_relative 'raml/trait'
 require_relative 'raml/exceptions'
 require_relative 'raml/include'
